% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_arlequin.R
\name{create_arlequin}
\alias{create_arlequin}
\title{Create Arlequin Input File (.arp)}
\source{
https://sites.google.com/site/thebantalab/tutorials?authuser=0#h.39zz6drphh3u
}
\usage{
create_arlequin(fasta, groups, output.dir = ".")
}
\arguments{
\item{fasta}{A DNAbin object representing the DNA alignment.}

\item{groups}{A data.frame with two columns: 'group' and 'name', representing
the sample groups.}

\item{output.dir}{A character string specifying the directory where the
output file will be saved. Default is the current working directory.}
}
\value{
This function does not return any value. It creates a file named 'output.arp'
in the specified directory.
}
\description{
This function generates an Arlequin input file (.arp) from a DNA alignment
and a groups data frame. The alignment should be in the DNAbin format, and
the groups data frame should have two columns: 'group' and 'name'. The function
checks the validity of the inputs and creates an Arlequin file in the specified
directory. If the output file already exists, the user will be prompted to delete it.
This script was adapted from Josh Banta (source)
}
\author{
Josh Banta
}
