% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hap_results.R
\name{hap_results}
\alias{hap_results}
\title{Format DNAsp Haplotype results}
\usage{
hap_results(vector)
}
\arguments{
\item{vector}{A character vector}
}
\value{
A tibble with two columns, haplotype and sequence
}
\description{
Splits a vector into a tibble with two columns, haplotype and sequence
}
\examples{
hap_results(vector)
}
