% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_genbank.R
\name{info_genbank}
\alias{info_genbank}
\title{Search for info from GenBank data}
\usage{
info_genbank(organism = "Aylacostoma[Organism]")
}
\arguments{
\item{organism}{The advanced search text from GenBank, like "Cichla piquiti\link{Organism}"}
}
\value{

}
\description{
Search for info from GenBank data
}
