% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptp_results.R
\name{ptp_results}
\alias{ptp_results}
\title{Format PTP Species Delimitation results}
\usage{
ptp_results(vector)
}
\arguments{
\item{vector}{}
}
\value{

}
\description{
Format PTP Species Delimitation results
}
