% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_seqs_by_hap.R
\name{select_seqs_by_hap}
\alias{select_seqs_by_hap}
\title{Select sequences by haplotype}
\usage{
select_seqs_by_hap(haps, seqs)
}
\arguments{
\item{haps}{A vector containing haplotype information.}

\item{seqs}{A DNA alignment object.}
}
\value{
A subset of sequences corresponding to the unique sequences by haplotype.
}
\description{
This function takes a set of haplotypes and a set of sequences, and returns a subset of sequences
corresponding to the unique haplotypes.
}
