% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_phy.R
\name{label_phy}
\alias{label_phy}
\title{Label phylip files}
\usage{
label_phy(seq_fasta, tree)
}
\arguments{
\item{seq_fasta}{The FASTA alignment file with full names}

\item{tree}{The .nwk file generated from .phy file}
}
\value{
a .nwk object
}
\description{
Label phylip files
}
\examples{
seq_normal <- ape::read.FASTA("alinhamento_cortado_2022_10_13_mioto_outgroup_3_limpo.fas")
tree_exe <- ape::read.tree("file.nwk")
new_tree <- label_phy(seq_normal, tree_exe)
write.tree(new_tree, "new_tree.nwk")
}
